+++ l2/agentDefsL2/agentDefsL2.ts
@@ -13,6 +13,7 @@ import {
 import { addD2Step, d2Result, updateD2Status } from '/_102020_/l2/agentDefsL2/helpers/d2Intents.js';
 import '/_102020_/l2/agentDefsL2/steps/entry10/agentD2Entry.js';
 import '/_102020_/l2/agentDefsL2/steps/input20/agentD2Input.js';
+import '/_102020_/l2/agentDefsL2/steps/contracts30/agentD2Contracts.js';
 
 export function createAgent(): IAgentAsync {
   return {